Add tests for mobile
